requirements.txt
PyMySQL==1.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
marshmallow==3.20.1
requests==2.33.1
pytest-mock==3.15.1
argparse==1.4.0
Flask==3.1.2
flask-cors>=3.0.10
mysql-connector-python==9.6.0
